Static folder; Sort of working deregister; Memes
Adds a static folder to the Spring Backend so that the Angular FrontEnd can work
properly.

Deregister only works when explicitly called by curl -x delete
[...]/shutdown. But it works. Sort of

I know uploading images to git is not a great idea, but it will ease
setup on another machine, and it's just 10 or so pics.